+++ RELEASE-NOTES.md
@@ -1,5 +1,6 @@
 ## 3.0.M1
 
+* #1 Support "recipientOverride" setting
 * #4 Upgrade to "simple-java-mail" 6.7.6
 
 ## 2.0.M1